Autosave on options page checkbox changes
I've had users complain that sync/welcome options aren't persisted.
I'm not sure if the options page is clear enough to users that they need to click save,
so I've set these to also automatically save on click. This mimics the auto-save of the
display settings in the customized apps page.