Made optional permissions more solid
Fixed up usability and workflow of granting/denying permissions.

A 'good' thing about permissions is that they need to be enabled by
a physical user interaction. A 'bad' thing about permissions is that they
apparently check the level of execution *or* asynchronous nature of the
physical event, so you can't just wrap the request in a promise down in an
ng-click event of a dynamically loaded HTML template. It breaks the good
design of AngularJS by forcing at least one chrome.* API to exist outside
of a service, but I'll deal.