Remove last remnant of requiring tabs permission
An earlier version of the extension required the tabs permission because
the tab id wasn't optional on chrome.tabs.update. This was left over in the
redirect.js file. I've removed that here so the tabs permission can also be optional.
This should make users feel much more comfortable with the required permissions.

I've also added an 'Options saved' message to indicate when quick save or save
has happened.