Fix cursor in address bar reported by many users
It seems half users want cursor in address bar while the other half
want it in the page. In an attempt to make debugging user issues with
cursor in address bar/in page, I changed to only  chrome.tabs.update.
In testing, I didn't notice any differences. I believe this is because
I tested in Chrome on a profile with a lot of apps/extensions.

To give both buckets of users the wanted functionality, I added an
update-only option which users can select to enable the new chrome.tabs.update
option. This should default to the original behavior for those users who
redirect to a search page and expect the cursor focus within the page.

Closes #56, #59

Pushing to github so I can test across systems more easily. Should be ready
by the end of today.