Apps page: fix uninstall and hrefs
Chromium was changed in revision 12d79a to require a user gesture for
chrome.management.uninstall.

Also indirectly reported in #70 is that all anchor tags on apps page display
app.appLaunchUrl appended relatively to New Tab Redirect's extension url. This
url doesn't matter because directives pull the evaluated url contents dynamically.
I've removed the standard hrefs from rollover icons on an app display, and changed the
displayed href for an application's launcher to display the correct target (this
follows the mantra of no user surprises). The click event bound to the element actually
prevents the default action and allows Chrome to launch the app by id. Doing this allows
the extension to launch apps and web pages consistently with little logic on my part.

Closes #70